merge: master <- development

+++ src/main/java/com/noroff/noroffassignment_7/repository/CharacterRepository.java
@@ -3,4 +3,7 @@ package com.noroff.noroffassignment_7.repository;
 import com.noroff.noroffassignment_7.model.Character;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CharacterRepository extends JpaRepository<Character, Integer> { }
+/**
+ * use: full CRUD in character with JpaRepository.
+ */
+public interface CharacterRepository extends JpaRepository<Character, Long> { }
