+++ src/main/java/com/noroff/noroffassignment_7/repository/MovieRepository.java
@@ -3,4 +3,7 @@ package com.noroff.noroffassignment_7.repository;
 import com.noroff.noroffassignment_7.model.Movie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MovieRepository extends JpaRepository<Movie, Integer> { }
+/**
+ * use: full CRUD in Movies with JpaRepository.
+ */
+public interface MovieRepository extends JpaRepository<Movie, Long> { }
